+++ src/gbtoolchainid.c
@@ -20,6 +20,7 @@
 #include "sig_gbnim.h"
 #include "sig_gbsdk.h"
 #include "sig_matlabgb.h"
+#include "sig_llvmgb_libgbxx.h"
 
 #include "sig_music.h"
 
@@ -214,6 +215,7 @@ void gbtools_detect(uint8_t * p_rom_data, uint32_t rom_size, bool strict_mode) {
     check_gbnim();
     check_gbsdk();
     check_matlabgb();
+    check_llvmgb_libgbxx();
 
     // Check for music drivers
     // May report multiple drivers (if found) in default output mode