+++ src/sig_llvmgb_libgbxx.h
@@ -1,0 +1,8 @@
+
+#ifndef _SIG_LLVMGB_LIBGBXX_H
+#define _SIG_LLVMGB_LIBGBXX_H
+
+void check_llvmgb_libgbxx(void);
+
+#endif // _SIG_LLVMGB_LIBGBXX_H
+